// src/components/Property/Property.types.ts
export type PropertyProps = {
  house: {
    id: number;
    name: string;
    tokenname: string;
    price: number;
    location: string;
    star: number;
    bedrooms: number;
    bathroom: number;
    meters: number;
    description: string;
    image: string;
    time: number;
    APY: number;
  };
};